move validate-token to server-common
