+++ apps/crn-server/test/routes/teams.route.test.ts
@@ -1,6 +1,6 @@
 import supertest from 'supertest';
 import Boom from '@hapi/boom';
-import { User } from '@asap-hub/auth';
+import { User, userMock } from '@asap-hub/auth';
 import { appFactory } from '../../src/app';
 import { FetchOptions } from '../../src/utils/types';
 import * as fixtures from '../fixtures/groups.fixtures';
@@ -11,7 +11,6 @@ import {
   getTeamResponse,
 } from '../fixtures/teams.fixtures';
 import { AuthHandler } from '../../src/middleware/auth-handler';
-import { userMock } from '../../src/utils/__mocks__/validate-token';
 import { FetchTeamsOptions } from '../../src/controllers/teams';
 
 describe('/teams/ route', () => {
