+++ apps/gp2-server/package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@asap-hub/auth": "workspace:*",
     "@asap-hub/model": "workspace:*",
+    "@asap-hub/server-common": "workspace:*",
     "@asap-hub/squidex": "workspace:*",
     "@graphql-typed-document-node/core": "^3.1.0",
     "@hapi/boom": "^9.1.2",
