+++ apps/gp2-server/test/mocks/auth-handler.mock.ts
@@ -1,5 +1,5 @@
+import { userMock } from '@asap-hub/auth';
 import { AuthHandler } from '../../src/middleware/auth-handler';
-import { userMock } from './validate-token.mock';
 
 export const authHandlerMock: AuthHandler = (req, _res, next) => {
   req.loggedInUser = userMock;
