+++ packages/auth/src/__mocks__/index.ts
@@ -1,7 +1,48 @@
+import { Auth0User, User } from '..';
+
 const config: typeof import('../config') = {
   domain: 'auth.example.com',
   clientID: 'client_id',
 };
 
 const auth0PubKeys = jest.requireActual('../pubKeys').default;
-export { config, auth0PubKeys };
+
+export const userMock: User = {
+  id: 'userMockId',
+  onboarded: true,
+  displayName: 'JT',
+  email: 'joao.tiago@asap.science',
+  firstName: 'Joao',
+  lastName: 'Tiago',
+  teams: [
+    {
+      id: 'team-id-1',
+      displayName: 'Awesome Team',
+      role: 'Project Manager',
+    },
+    {
+      id: 'team-id-3',
+      displayName: 'Zac Torres',
+      role: 'Collaborating PI',
+    },
+  ],
+  algoliaApiKey: 'test-mock-key',
+};
+
+const getAuth0UserMock = ({ origin }: { origin: string }): Auth0User => ({
+  [`${origin}/user`]: userMock,
+  given_name: 'Joao',
+  family_name: 'Tiago',
+  nickname: 'joao.tiago',
+  name: 'Joao Tiago',
+  picture: 'https://lh3.googleusercontent.com/awesomePic',
+  locale: 'en',
+  updated_at: '2020-10-27T17:55:23.418Z',
+  email: 'joao.tiago@asap.science',
+  iss: 'https://asap-hub.us.auth0.com/',
+  sub: 'google-oauth2|awesomeGoogleCode',
+  aud: 'audience',
+  nonce: 'onlyOnce',
+});
+
+export { config, auth0PubKeys, getAuth0UserMock };
