+++ packages/auth/src/index.ts
@@ -46,3 +46,5 @@ export type Auth0 = {
   | 'getTokenWithPopup'
   | 'logout'
 >;
+
+export * from './__mocks__';
