+++ packages/server-common/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@asap-hub/server-common",
+  "private": true,
+  "version": "0.1.0",
+  "engines": {
+    "node": "14.15.x"
+  },
+  "module": "build/index.js",
+  "main": "build-cjs/index.js",
+  "types": "build/index.d.ts",
+  "scripts": {
+    "build:babel": "../../scripts/build-babel.sh build",
+    "watch:babel": "../../scripts/build-babel.sh watch"
+  },
+  "dependencies": {
+    "@asap-hub/auth": "workspace:*",
+    "@babel/runtime-corejs3": "^7.14.0",
+    "jsonwebtoken": "^8.5.1"
+  },
+  "devDependencies": {
+    "@asap-hub/eslint-config-asap-hub": "workspace:*",
+    "@types/jest": "^26.0.23",
+    "@types/jsonwebtoken": "^8.5.1"
+  }
+}
