+++ packages/server-common/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "build"
+  },
+  "include": ["src", "types"],
+  "references": [{ "path": "../../packages/auth" }]
+}
